Tidy coins and explorer code
