+++ cryptos/coins/bitcoin_cash.py
@@ -1,13 +1,14 @@
-from .bitcoin import Bitcoin
+from .base import BaseCoin
 from ..transaction import SIGHASH_ALL, SIGHASH_FORKID
 from ..explorers import blockdozer
 
-class BitcoinCash(Bitcoin):
+class BitcoinCash(BaseCoin):
     coin_symbol = "bcc"
     display_name = "Bitcoin Cash"
     segwit_supported = False
     magicbyte = 0
     script_magicbyte = 5
+    explorer = blockdozer
     hashcode = SIGHASH_ALL + SIGHASH_FORKID
     testnet_overrides = {
         'display_name': "Bitcoin Cash Testnet",
@@ -15,18 +16,3 @@ class BitcoinCash(Bitcoin):
         'magicbyte': 111,
         'script_magicbyte': 196
     }
-
-    def unspent(self, *addrs):
-        return blockdozer.unspent(*addrs, coin_symbol=self.coin_symbol)
-
-    def history(self, *addrs):
-        return blockdozer.history(*addrs, coin_symbol=self.coin_symbol)
-
-    def fetchtx(self, tx):
-        return blockdozer.fetchtx(tx, coin_symbol=self.coin_symbol)
-
-    def txinputs(self, tx):
-        return blockdozer.txinputs(tx, coin_symbol=self.coin_symbol)
-
-    def pushtx(self, tx):
-        return blockdozer.pushtx(tx, coin_symbol=self.coin_symbol)
