+++ cryptos/coins/dogecoin.py
@@ -1,4 +1,5 @@
 from .bitcoin import Bitcoin
+from ..explorers import sochain
 
 
 class Doge(Bitcoin):
@@ -6,6 +7,7 @@ class Doge(Bitcoin):
     display_name = "Dogecoin"
     segwit_supported = False
     magicbyte = 30
+    explorer = sochain
     testnet_overrides = {
         'display_name': "Dogecoin Testnet",
         'coin_symbol': "Dogecoin",
