+++ cryptos/explorers/dash_siampm.py
@@ -1,26 +1,26 @@
 from . import base_insight as insight
 
-def get_url(is_test):
-    if is_test:
-        return "https://test.insight.dash.siampm.com/api"
-    return "https://insight.dash.siampm.com/api"
+def get_url(coin_symbol):
+    if coin_symbol == "DASH":
+        return "https://insight.dash.siampm.com/api"
+    return "https://test.insight.dash.siampm.com/api"
 
-def unspent(*args, testnet=False):
-    base_url = get_url(testnet)
+def unspent(*args, coin_symbol="DASH"):
+    base_url = get_url(coin_symbol)
     return insight.unspent(base_url, *args)
 
-def fetchtx(tx, testnet=False):
-    base_url = get_url(testnet)
+def fetchtx(tx, coin_symbol="DASH"):
+    base_url = get_url(coin_symbol)
     return insight.fetchtx(base_url, tx)
 
-def txinputs(tx, testnet=False):
-    base_url = get_url(testnet)
+def txinputs(tx, coin_symbol="DASH"):
+    base_url = get_url(coin_symbol)
     return insight.txinputs(base_url, tx)
 
-def pushtx(tx, testnet=False):
-    base_url = get_url(testnet)
-    return insight.pushtx(base_url, "DASHTEST" if testnet else "DASH", tx)
+def pushtx(tx, coin_symbol="DASH"):
+    base_url = get_url(coin_symbol)
+    return insight.pushtx(base_url, coin_symbol, tx)
 
-def history(*args,  testnet=False):
-    base_url = get_url(testnet)
+def history(*args,  coin_symbol="DASH"):
+    base_url = get_url(coin_symbol)
     return insight.history(base_url, *args)
